unnötige Klasse entfernt + RolesAllowed zu den exercise Views hinzugefügt
